refactor: DRY up showcases with a helper
