+++ packages/key-facts/key-facts.showcase.js
@@ -1,27 +1,10 @@
 /* eslint-disable react/prop-types */
 import React from "react";
-import invert from "lodash.invert";
 import { ScrollView } from "react-native";
-import Context from "@times-components/context";
 import { LateralSpacingDecorator } from "@times-components/storybook";
-import { colours, scales } from "@times-components/styleguide";
-import KeyFacts from "./src/key-facts";
 import data from "./fixtures/key-facts-showcase.json";
 import dataNoTitle from "./fixtures/key-facts-no-title-showcase.json";
-
-const selectSection = select =>
-  select("Section", invert(colours.section), colours.section.default);
-const selectScales = select => select("Scale", scales, scales.medium);
-
-const renderKeyFacts = (ast, select) => {
-  const scale = selectScales(select);
-  const sectionColour = selectSection(select);
-  return (
-    <Context.Provider value={{ theme: { scale, sectionColour } }}>
-      <KeyFacts ast={ast} onLinkPress={() => {}} />
-    </Context.Provider>
-  );
-};
+import renderKeyFacts from "./showcase-helper";
 
 export default {
   children: [
@@ -32,7 +15,7 @@ export default {
     {
       component: ({ select }) => (
         <ScrollView style={{ width: "100%" }}>
-          {renderKeyFacts(data, select)}
+          {renderKeyFacts({ ast: data, hasScaling: true, select })}
         </ScrollView>
       ),
       name: "with title",
@@ -42,7 +25,7 @@ export default {
     {
       component: ({ select }) => (
         <ScrollView style={{ width: "100%" }}>
-          {renderKeyFacts(dataNoTitle, select)}
+          {renderKeyFacts({ ast: dataNoTitle, hasScaling: true, select })}
         </ScrollView>
       ),
       name: "without title",
